+++ src/components/MapPolygon.jsx
@@ -2,9 +2,7 @@ import React, { useState, useCallback } from 'react';
 import { GoogleMap, Marker, useJsApiLoader } from '@react-google-maps/api';
 import Popover from '@mui/material/Popover';
 import Button from '@mui/material/Button';
-import EventForm from './EventForm'
-
-import {postData} from "../services/service";
+import EventForm from './EventForm';
 
 const containerStyle = {
     width: '100%',
@@ -49,28 +47,7 @@ const MapPolygon = () => {
         setAnchorEl({top: event.domEvent.clientY, left: event.domEvent.clientX});
         setOpenPopover(true);
     };
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-
-        const data = {
-            coordinates: [[...markerPosition.lat, markerPosition.lng]],
-            eventType,
-            eventDate,
-            description
-        }
-
-        postData(`https://capdragons.koyeb.app/api/v1/events`, JSON.stringify(data))
-            .then(res => {
-                const {status} = res;
-                if (status.toString().startsWith("4")) {
-                    console.log("error")
-                } else {
-                   console.log("POSTED!")
-                }
-            })
-            .catch(e => console.log(e));
-    }
+    
 
     const onLoad = useCallback(function callback(mapInstance) {
         mapInstance.setZoom(defaultZoomLevel);
@@ -114,7 +91,7 @@ const MapPolygon = () => {
             }}
         >
             <Button onClick={handleClose}>Close me!</Button>
-            <EventForm coordinates={{lat: null, lng: null}} />
+            <EventForm coordinates={markerPosition} />
         </Popover>
         </>
     ) : <></>;
